Finished updating to use typescript

+++ src/routes/api/forum/post.ts
@@ -1,10 +1,11 @@
 import { Router } from 'express';
 import { Request, Response } from 'express';
+import mongoose from 'mongoose';
 import * as log from '../../../logging/logging';
 import authorization from '../../../middleware/api/authorization';
 import Cache from '../../../middleware/api/Cache';
 import updateLastOnline from '../../../middleware/api/updateLastOnline';
-import Category from '../../../models/Category';
+import Category, { ICategory } from '../../../models/Category';
 import Post from '../../../models/Post';
 import Topic, { ITopic } from '../../../models/Topic';
 import User, { IUser } from '../../../models/User';
@@ -12,6 +13,24 @@ import User, { IUser } from '../../../models/User';
 const router = Router();
 router.use('/', authorization);
 router.use('/', updateLastOnline);
+
+const userIsMod = async (
+  user: IUser,
+  cat: string | mongoose.Schema.Types.ObjectId
+): Promise<boolean> => {
+  return Category.findById(cat)
+    .then(c => {
+      if (c) {
+        return user._id in c.moderators;
+      } else {
+        return false;
+      }
+    })
+    .catch((e: Error) => {
+      log.error(e.message);
+      return false;
+    });
+};
 
 // @route   GET /forum/post
 // @desc    Returns a short list of tpics
@@ -77,55 +96,58 @@ router.post(
   async (req: Request, res: Response): Promise<void> => {
     log.trace('POST /forum/post reached endpoint');
     /* Pull in all relevant data points */
-    let topic: ITopic | undefined;
-    let user: IUser | undefined;
-    try {
-      topic = await Topic.findById(req.body.topic).then(t => {
-        if (t) {
-          return t;
-        } else {
-          res.status(404).json({
-            message: 'Resource was not found on this server',
-            success: false
-          });
-        }
+    let topic: ITopic;
+    let user: IUser;
+
+    const tempUser = await User.findOne({
+      session_token: req.get('authorization')
+    });
+
+    if (!tempUser) {
+      res.status(404).json({
+        message: 'The requested resource was not found on the server',
+        success: false
       });
-      user = await User.findOne({
-        session_token: req.get('authorization')
-      }).then(t => {
-        if (t) {
-          return t;
-        } else {
-          res.status(404).json({
-            message: 'Resource was not found on this server',
-            success: false
-          });
-        }
-      });
-    } catch (e) {
-      log.error(e.message);
-      res
-        .status(500)
-        .json({ success: false, message: 'An unknown error has occured' });
+      return;
+    } else {
+      user = tempUser;
     }
-    if (topic) {
-      new Post({
-        category: topic.category,
-        message: req.body.message,
-        topicId: topic._id,
-        userId: user._id
-      })
-        .save()
-        .then(post =>
-          res.json({ success: true, message: 'Posted successfully', post })
-        )
-        .catch(e => {
-          log.error(e);
-          res
-            .status(500)
-            .json({ success: false, message: 'An unknown error has occured' });
+
+    const tempTopic = await Topic.findById(req.body.topic);
+
+    if (!tempTopic) {
+      res.status(404).json({
+        message: 'The requested resource was not found on the server',
+        success: false
+      });
+      return;
+    } else {
+      topic = tempTopic;
+    }
+
+    Topic.findById(req.body.topic);
+
+    new Post({
+      category: topic.category,
+      message: req.body.message,
+      poster: user._id,
+      topicId: topic._id
+    })
+      .save()
+      .then(post =>
+        res.json({
+          message: 'Posted successfully',
+          post,
+          success: true
+        })
+      )
+      .catch(e => {
+        log.error(e);
+        res.status(500).json({
+          message: 'An unknown error has occured',
+          success: false
         });
-    }
+      });
   }
 );
 
@@ -141,14 +163,13 @@ router.get('/:id', Cache.retrieve, (req: Request, res: Response): void => {
           message: 'Resource was not found',
           sucess: false
         });
-        return;
+      } else {
+        Cache.cache(60)(req, res, {
+          message: 'Post retrieved successfully',
+          post,
+          success: true
+        });
       }
-
-      Cache.cache(60)(req, res, {
-        message: 'Post retrieved successfully',
-        post,
-        success: true
-      });
     })
     .catch(e => {
       res
@@ -166,47 +187,55 @@ router.delete(
     log.trace(`DELETE /forum/post/${req.params.id} reached endpoint`);
     const token = req.get('authorization');
     let user: IUser;
-    try {
-      user = await User.findOne({ session_token: token });
-    } catch (e) {
-      res
-        .status(500)
-        .json({ success: false, message: 'An unknown error has occured' });
-    }
+
+    User.findOne({ session_token: token })
+      .then(u => {
+        if (u) {
+          user = u;
+        } else {
+          res.status(404).json({
+            message: 'The requested resource was not found on the server',
+            success: false
+          });
+        }
+      })
+      .catch((e: Error) => {
+        log.error(e.message);
+        res
+          .status(500)
+          .json({ message: 'An unknown error occured', success: false });
+      });
+
     Post.findById(req.params.id)
       .then(async post => {
         let canDelete: boolean;
         /* Check validation. Can only be done by moderators of the category, admins and the user that made the post */
-        try {
+        if (post) {
           canDelete =
             user.role === 'admin' ||
             post.poster === user._id ||
-            user._id in (await Category.findById(post.category)).moderators;
-        } catch (e) {
-          res
-            .status(500)
-            .json({ success: false, message: 'An unknown error has occured' });
-        }
+            (await userIsMod(user, post.category));
 
-        if (!canDelete) {
-          res.status(403).json({
-            message: 'You do not have authorization to perform this action',
-            success: false
-          });
-        } else {
-          Post.findByIdAndDelete(req.params.id)
-            .then(() => {
-              res.json({
-                message: 'Post deleted successfully',
-                success: true
-              });
-            })
-            .catch((e: Error) => {
-              res.status(500).json({
-                message: 'An unknown error occured',
-                success: false
-              });
+          if (!canDelete) {
+            res.status(403).json({
+              message: 'You do not have authorization to perform this action',
+              success: false
             });
+          } else {
+            Post.findByIdAndDelete(req.params.id)
+              .then(() => {
+                res.json({
+                  message: 'Post deleted successfully',
+                  success: true
+                });
+              })
+              .catch((e: Error) => {
+                res.status(500).json({
+                  message: 'An unknown error occured',
+                  success: false
+                });
+              });
+          }
         }
       })
       .catch((e: Error) => {
@@ -227,15 +256,25 @@ router.put(
     log.trace(`PUT /forum/post/${req.params.id} reached endpoint`);
     const token = req.get('authorization');
     let user: IUser;
-    try {
-      user = await User.findOne({ session_token: token });
-    } catch (e) {
-      res.status(500).json({
-        message: 'An unknown error has occured',
-        success: false
+
+    User.findOne({ session_token: token })
+      .then(u => {
+        if (u) {
+          user = u;
+        } else {
+          res.status(404).json({
+            message: 'The requested resource was not found on the server',
+            success: false
+          });
+        }
+      })
+      .catch((e: Error) => {
+        log.error(e.message);
+        res
+          .status(500)
+          .json({ message: 'An unknown error occured', success: false });
       });
-      return;
-    }
+
     Post.findById(req.params.id)
       .then(async post => {
         if (!post) {
@@ -262,8 +301,7 @@ router.put(
           const canUpdate: boolean =
             user.role === 'admin' ||
             post.poster === user._id ||
-            user._id in
-              (await Category.findOne({ _id: post.category })).moderators;
+            (await userIsMod(user, post.category));
           update(canUpdate);
         } catch (e) {
           res.status(500).json({
